fix warning

+++ src/controls/reactive_fallback.cpp
@@ -28,7 +28,7 @@ NodeStatus ReactiveFallback::tick()
         {
             case NodeStatus::RUNNING:
             {
-                for(int i=index+1; i < childrenCount(); i++)
+                for(size_t i=index+1; i < childrenCount(); i++)
                 {
                     haltChild(i);
                 }
